+++ src/extension.ts
@@ -1,22 +1,7 @@
 'use strict';
-import { ExtensionContext, commands, window, EndOfLine, Position, workspace, TextEditorEdit, Range } from 'vscode';
+import { ExtensionContext, commands, window, Position, workspace, TextEditorEdit, Range, TextLine } from 'vscode';
 
 export function activate(context: ExtensionContext) {
-    context.subscriptions.push(commands.registerTextEditorCommand('extension.drawBox', (textEditor, edit) => {
-        if (textEditor.document.languageId !== 'markdown' || textEditor.selections.length !== 1) {
-            window.showErrorMessage('Must be a MarkDown editor with a sole selection!');
-            return;
-        }
-
-        const { start: startPosition, end: endPosition } = textEditor.selection;
-        const { eol, getText, offsetAt } = textEditor.document;
-        const startOffset = offsetAt(startPosition);
-        const endOffset = offsetAt(endPosition);
-        const { style } = workspace.getConfiguration('boxDrawing');
-        const { replacementText } = drawBox(startPosition, endPosition, eol, getText(), startOffset, endOffset, style);
-        edit.replace(textEditor.selection, replacementText);
-    }));
-
     context.subscriptions.push(commands.registerTextEditorCommand('extension.insertCanvas', (textEditor, edit) => {
         if (textEditor.document.languageId !== 'markdown') {
             window.showErrorMessage('Must be a MarkDown editor!');
@@ -25,6 +10,17 @@ export function activate(context: ExtensionContext) {
 
         // TODO: https://github.com/Microsoft/vscode/issues/14756
         edit.replace(textEditor.selection, `\n${Array(30).fill(' '.repeat(120)).join('\n')}\n`);
+    }));
+
+    context.subscriptions.push(commands.registerTextEditorCommand('extension.drawBox', (textEditor, edit) => {
+        if (textEditor.document.languageId !== 'markdown' || textEditor.selections.length !== 1) {
+            window.showErrorMessage('Must be a MarkDown editor with a sole selection!');
+            return;
+        }
+
+        const { start: startPosition, end: endPosition } = textEditor.selection;
+        const { style } = workspace.getConfiguration('boxDrawing');
+        drawBox(startPosition, endPosition, edit, textEditor.document.lineAt, style);
     }));
 
     context.subscriptions.push(commands.registerTextEditorCommand('extension.drawArrow', (textEditor, edit) => {
@@ -38,77 +34,49 @@ export function activate(context: ExtensionContext) {
     }));
 }
 
-export function drawBox(startPosition: Position, endPosition: Position, eol: EndOfLine, text: string, startOffset: number, endOffset: number, style: 'ascii' | 'unicode') {
+// TODO: Respect the switch for ASCII/Unicode
+export function drawBox(startPosition: Position, endPosition: Position, edit: TextEditorEdit, lineAt: (index: number) => TextLine, style: 'ascii' | 'unicode') {
     const { character: startCharacter, line: startLine } = startPosition;
     const { character: endCharacter, line: endLine } = endPosition;
 
-    const minCharacter = Math.min(startCharacter, endCharacter);
-    const maxCharacter = Math.max(startCharacter, endCharacter);
-    const boxWidth = maxCharacter - minCharacter;
+    const nearCharacter = Math.min(startCharacter, endCharacter);
+    const farCharacter = Math.max(startCharacter, endCharacter);
+    const nearLine = Math.min(startLine, endLine);
+    const farLine = Math.max(startLine, endLine);
 
-    const minLine = Math.min(startLine, endLine);
-    const maxLine = Math.max(startLine, endLine);
-    const boxHeight = maxLine - minLine + 1;
+    for (let index = nearLine; index <= farLine; index++) {
+        const line = lineAt(index);
 
-    let eolText = '';
-    switch (eol) {
-        case EndOfLine.LF: eolText = '\n'; break;
-        case EndOfLine.CRLF: eolText = '\r\n'; break;
-        default: throw new Error('Unexpected EOL value.');
-    }
-
-    const selectionText = text.substring(startOffset, endOffset);
-
-    let replacementText = '';
-
-    let cursor = selectionText.indexOf(eolText) + eolText.length;
-
-    replacementText += '┏';
-    replacementText += '━'.repeat(boxWidth - 2); // Box top side
-    replacementText += '┓';
-    if (boxWidth > cursor) {
-        replacementText += selectionText.substring(cursor - eolText.length, cursor); // Rest of line
-    } else {
-        replacementText += selectionText.substring(boxWidth, cursor); // Rest of line
-    }
-
-    for (let line = 1; line < boxHeight; line++) {
-        let leftOffset = cursor + minCharacter;
-        let rightOffset = leftOffset + boxWidth;
-        replacementText += selectionText.substring(cursor, leftOffset); // Before box line part
-        cursor = selectionText.indexOf(eolText, cursor) + eolText.length;
-
-        const isLast = line === boxHeight - 1;
-        if (!isLast) {
-            replacementText += '┃'; // Box left side
-            // Box line part
-            if (rightOffset > cursor) {
-                const boxLine = selectionText.substring(leftOffset + 1, cursor - 1);
-                // TODO: Fix this, it's weird, this code must be better generalizeable
-                if (boxLine !== eolText) {
-                    replacementText += boxLine.padEnd(boxWidth - 2, ' ');
-                } else {
-                    replacementText += ' '.repeat(boxWidth - 2);
-                }
-
-                replacementText += '┃'; // Box right side
-                replacementText += selectionText.substring(cursor - eolText.length, cursor); // After box line part
+        const nearSymbol = line.lineNumber === nearLine ? '┏' : (line.lineNumber === farLine ? '┗' : '┃');
+        const farSymbol = line.lineNumber === nearLine ? '┓' : (line.lineNumber === farLine ? '┛' : '┃');
+        if (line.range.end.character < nearCharacter) {
+            // Pad the line long enough to reach the box
+            edit.insert(line.range.end, ' '.repeat(nearCharacter - line.range.end.character));
+            const position = new Position(line.lineNumber, nearCharacter);
+            if (line.lineNumber !== nearLine && line.lineNumber !== farLine) {
+                edit.insert(position, '┃' + ' '.repeat(farCharacter - nearCharacter - 2) + '┃');
             } else {
-                replacementText += selectionText.substring(leftOffset + 1, rightOffset - 1);
-                replacementText += '┃'; // Box right side
-                replacementText += selectionText.substring(rightOffset, cursor); // After box line part
+                edit.insert(position, nearSymbol + '━'.repeat(farCharacter - nearCharacter - 2) + farSymbol);
+            }
+        } else if (line.range.end.character < farCharacter) {
+            // End line at the box for simple insertion
+            edit.delete(new Range(new Position(line.lineNumber, nearCharacter), line.range.end));
+            const position = line.range.end;
+            if (line.lineNumber !== nearLine && line.lineNumber !== farLine) {
+                window.showInformationMessage((farCharacter - line.range.end.character).toString());
+                edit.insert(position, nearSymbol + line.text.substring(nearCharacter + 1) + ' '.repeat(farCharacter - line.range.end.character - 1) + farSymbol);
+            } else {
+                edit.insert(position, nearSymbol + '━'.repeat(farCharacter - nearCharacter - 2) + farSymbol);
             }
         } else {
-            replacementText += '┗'; // Box left side
-            replacementText += '━'.repeat(boxWidth - 2); // Box line part
-            replacementText += '┛'; // Box right side
-            if (cursor - eolText.length !== -1) {
-                replacementText += selectionText.substring(cursor - eolText.length);
+            const range = new Range(new Position(line.lineNumber, nearCharacter), new Position(line.lineNumber, farCharacter));
+            if (line.lineNumber !== nearLine && line.lineNumber !== farLine) {
+                edit.replace(range, nearSymbol + line.text.substring(nearCharacter + 1, farCharacter - 1) + farSymbol);
+            } else {
+                edit.replace(range, nearSymbol + '━'.repeat(farCharacter - nearCharacter - 2) + farSymbol);
             }
         }
     }
-
-    return { replacementText, boxHeight, boxWidth };
 }
 
 export function drawArrow(startPosition: Position, endPosition: Position, edit: TextEditorEdit) {