+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Respect the `boxDrawing.style` configuration setting when drawing boxes
 - Use Unicode middle dot instead of a dash when drawing arrows
+- Switch to a new box drawing algorithm
 
 ## `3.0.0` (2018-05-09)